==> code/03-phyluce.Rmd <==
---
title: "Harvesting UCEs from Genomes"
author: "Megan Ewing"
date: "2025-07-24"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Harvesting UCE Loci

This is following [phyluce tutorial III: harvesting UCE loci from genomes](https://phyluce.readthedocs.io/en/latest/tutorials/tutorial-3.html).

Important folder paths:

-   genomes: `../data/genomes_refseq/ncbi_dataset/data/`

genomes inlcude the **17 bivalve genomes** and the **3 molluscan outgroup genomes**

Necessary packages: [fatotwobit](https://anaconda.org/bioconda/ucsc-fatotwobit) , [twobitinfo](https://anaconda.org/bioconda/ucsc-twobitinfo), and [twobittofa](https://anaconda.org/bioconda/ucsc-twobittofa) (which i have in an enviornment 'twoBit' for):

```{bash}

### to open the twoBit environment where the packages live: 
# 
# conda activate twoBit
# 
### if conda activate gives an error like command not found, you'll need to run:
# 
# eval "$(/opt/anaconda/anaconda3/bin/conda shell.bash hook)"
# 
### and then 
# 
# conda activate twoBit
#
#
### don't forget to `conda deactivate` at end of session!
```

Need to move the genomes to the phyluce data folder as 2bits -- do this as loop. also get info

### TwoBit Genomes

```{bash}

# Input and output base directories
INPUT_BASE="../data/genomes_refseq/ncbi_dataset/data"
OUTPUT_BASE="../data/phyluce"

# Loop over each accession-named folder in the input directory
for acc_dir in "$INPUT_BASE"/*; do
  # Get the accession ID (folder name)
  accession=$(basename "$acc_dir")

  # Define input and output file paths
  fna_file="$acc_dir/cds_from_genomic.fna"
  out_dir="$OUTPUT_BASE/$accession"
  two_bit_file="$out_dir/${accession}.2bit"
  sizes_file="$out_dir/sizes.tab"

  # Check if .fna file exists
  if [[ ! -f "$fna_file" ]]; then
    echo "No .fna file found for $accession, skipping."
    continue
  fi

  # Create output directory if needed
  mkdir -p "$out_dir"

  # Convert .fna to .2bit
  echo "Converting $accession -> .2bit"
  /opt/anaconda/anaconda3/envs/twoBit/bin/faToTwoBit "$fna_file" "$two_bit_file"

  # Get sizes info
  /opt/anaconda/anaconda3/envs/twoBit/bin/twoBitInfo "$two_bit_file" "$sizes_file"

  # Preview sizes file
  echo "Preview of $sizes_file:"
  head -n 5 "$sizes_file"
  echo "-------------------------------------"
done


```

### Align the probes to the genomes

"Now, we need to think of some name for the database to create (here \`tutorial3.sqlite\`), the name of an output in which to store the \`lastz\` search results, the path to the genome sequences, the name of the genome sequences, the path to the probe file, and a number of compute cores to use:" (from phyluce tutorial)

currently, the folders and files within are `GCF_#########.#`, but the `.#` will confuse phyluce, so let's get rid of those.

```{r}
library(stringr)
library(fs)

base_dir <- "../data/phyluce"

# rename directories
dirs <- dir_ls(base_dir, type = "directory", recurse = FALSE, regexp = "^.*/GCF_.*\\.\\d+$")
for (dir in dirs) {
  dir_name <- path_file(dir)
  new_dir_name <- str_replace(dir_name, "\\.\\d+$", "")
  new_dir_path <- fs::path(fs::path_dir(dir), new_dir_name)
  file_move(dir, new_dir_path)  # <- Corrected line
  cat("Renamed dir:", dir, "->", new_dir_path, "\n")
}


# rename 2bit files
# List all .2bit files under subdirectories (recursive = TRUE)
bit_files <- list.files(path = base_dir, pattern = "^GCF_\\d+\\.\\d+\\.2bit$", 
                        recursive = TRUE, full.names = TRUE)

for (old_path in bit_files) {
  # Extract current filename
  filename <- basename(old_path)
  
  # Remove the dot-version (e.g., .1) from filename
  new_filename <- sub("(GCF_\\d+)\\.\\d+\\.2bit", "\\1.2bit", filename)
  
  if (new_filename != filename) {
    # Build new path in same directory
    dir_path <- dirname(old_path)
    new_path <- file.path(dir_path, new_filename)
    
    # Rename the file
    file.rename(old_path, new_path)
    cat("Renamed:", old_path, "->", new_path, "\n")
  }
}
```

```{bash}
# list directory contents to check
# reason for the extra ls commands is to have it print in a way that can just be copied and pasted as the scaffold in the probe align chunk
ls -m ../data/phyluce/ | tr -d ','
```

We also need to adjust the sequence names within the .2bit files. now, this may seem a little two-steps-forward-one-step-back, but to do this we'll convert the .2bit files back to fasta, change the sequence names, and then convert back to .2bit and delete the intermediate files. The reason we didn't just change the fastas in the first place, is because we want to keep those exactly how they came from ncbi. we just want some change in the .2bit sequence names so that we can run lastz without the names getting overly truncated.

```{bash}

# Base directory where .2bit files live
BASE_DIR="../data/phyluce"

# Loop through all .2bit files
find "$BASE_DIR" -name "*.2bit" | while read -r twoBitFile; do
    # Extract directory and base name
    dir=$(dirname "$twoBitFile")
    base=$(basename "$twoBitFile" .2bit)

    echo "Processing: $twoBitFile"

    # Convert to .fasta
    /opt/anaconda/anaconda3/envs/twoBit/bin/twoBitToFa "$twoBitFile" "$dir/${base}.fasta"

    # Remove lcl| from headers
    sed 's/^>lcl|/>/' "$dir/${base}.fasta" > "$dir/${base}_clean.fasta"

    # Convert back to .2bit
    /opt/anaconda/anaconda3/envs/twoBit/bin/faToTwoBit "$dir/${base}_clean.fasta" "$dir/${base}_clean.2bit"

    echo "Cleaned .2bit written: $dir/${base}_clean.2bit"
done

# Replace original .2bit files with cleaned versions
find "$BASE_DIR" -name "*_clean.2bit" | while read -r cleanBit; do
    original="${cleanBit%_clean.2bit}.2bit"
    mv "$cleanBit" "$original"
done

# Optionally remove intermediate FASTAs
find "$BASE_DIR" -name "*.fasta" -delete

```

### Bait Sequences for UCE

Bait sequences were provided courtesy of Yi-Xuan Li, of Hong Kong Baptist University, from [Yi-Xuan et al. (2024)](https://academic.oup.com/sysbio/article-abstract/74/1/16/7758682)

you'll need to ensure [phyluce](https://github.com/faircloth-lab/phyluce/releases/tag/v1.7.3) is installed for this

phyluce installed mamba environment: `/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce/`

```{bash}
## info to initialize mamba / phyluce in terminal : 
#
# conda create -n phyluce-py3 python=3.8 phyluce -c bioconda -c conda-forge
#
# eval "$(/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda shell.bash hook)"
# eval "$(mamba shell hook --shell bash)"
#
## activates base 
# mamba activate
#
## activates environment i created
# mamba activate phyluce 
```

scaffold lists below are just the copied and pasted output of directory contents printed 2 chunks ago

```{bash}
# for v1 baits

cd ../data/phyluce/

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_probe_run_multiple_lastzs_sqlite \
        --db uce_frombaits_v1 \
        --output v1-genome-lastz \
        --scaffoldlist GCF_000327385 GCF_002022765 GCF_002113885 GCF_020536995 GCF_021730395 GCF_021869535 GCF_025612915 GCF_026571515 GCF_026914265 GCF_031769215 GCF_032062105 GCF_033153115 GCF_036588685 GCF_041381155 GCF_902652985 GCF_947568905 GCF_963676685 GCF_963853765 \
        --genome-base-path ./ \
        --probefile Baits_Bivalve_v1.fasta \
        --cores 10 \
        --identity 80 \
        2> phyluce.log
        
```

```{bash}

# for v2 baits

cd ../data/phyluce/

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_probe_run_multiple_lastzs_sqlite \
        --db uce_frombaits_v2 \
        --output v2-genome-lastz \
        --scaffoldlist GCF_000327385 GCF_002022765 GCF_002113885 GCF_020536995 GCF_021730395 GCF_021869535 GCF_025612915 GCF_026571515 GCF_026914265 GCF_031769215 GCF_032062105 GCF_033153115 GCF_036588685 GCF_041381155 GCF_902652985 GCF_947568905 GCF_963676685 GCF_963853765 \
        --genome-base-path ./ \
        --probefile Baits_Bivalve_v2.fasta \
        --cores 10 \
        --identity 80 \
        2> phyluce_v2.log

```

### Extracting FASTA sequence matching UCE loci from genome sequences

need to create a configuration file that tells the program where to find each genome

```{r}
library(fs)
library(stringr)

# Set paths
base_dir <- path_expand("~/bivalve_gametogenesis/data/phyluce")
output_file <- path_expand("~/bivalve_gametogenesis/data/phyluce/genomes.conf")  

# List all .2bit files recursively in GCF_* directories
bit_files <- dir_ls(base_dir, recurse = TRUE, glob = "*.2bit")

# Filter for GCF-named .2bit files
bit_files <- bit_files[grepl("GCF_\\d+/", bit_files)]

# Build config lines
config_lines <- c("[scaffolds]")

for (file_path in bit_files) {
  file_name <- path_file(file_path)
  gcf_id <- str_match(file_name, "^(GCF_\\d+)")[,2]
  full_path <- path_expand(file_path)
  config_lines <- c(config_lines, paste0(gcf_id, ":", full_path))
}

# Write to genomes.conf
writeLines(config_lines, output_file)
```

```{bash}
head ../data/phyluce/genomes.conf
```

```{bash}

# for v1

cd ../data/phyluce/

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_probe_slice_sequence_from_genomes \
        --lastz v1-genome-lastz \
        --conf genomes.conf \
        --flank 500 \
        --name-pattern "Baits_Bivalve_v1.fasta_v_{}.lastz.clean" \
        --output v1-genome-fasta
```

```{bash}
# for v2

cd ../data/phyluce/

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_probe_slice_sequence_from_genomes \
        --lastz v2-genome-lastz \
        --conf genomes.conf \
        --flank 500 \
        --name-pattern "Baits_Bivalve_v2.fasta_v_{}.lastz.clean" \
        --output v2-genome-fasta \

```

below creates a summary table showing how many sequences hit for each step. notes that the fasta sequence \# may be lower than the lastz \# due to sequences getting cutoff in the flanking region

```{r}
library(tidyverse)

# Define directories
base_dir <- "../data/phyluce"
paths <- list(
  v1_lastz = file.path(base_dir, "v1-genome-lastz"),
  v2_lastz = file.path(base_dir, "v2-genome-lastz"),
  v1_fasta = file.path(base_dir, "v1-genome-fasta"),
  v2_fasta = file.path(base_dir, "v2-genome-fasta")
)

# Function to count lines (lastz) or headers (fasta)
count_lastz <- function(file) {
  # Count non-empty lines = one match per line
  length(readLines(file))
}

count_fasta <- function(file) {
  # Count how many lines start with >
  sum(readLines(file) %>% str_starts(">"))
}

# Parse sample ID from filenames
parse_sample_lastz <- function(filename) {
  str_extract(filename, "(?<=_v_GCF_)\\d+")
}

parse_sample_fasta <- function(filename) {
  str_extract(tolower(filename), "(?<=gcf_)\\d+")
}

# Collect counts
collect_counts <- function(folder, type) {
  files <- list.files(folder, full.names = TRUE)
  if (length(files) == 0) return(data.frame())
  
  tibble(
    file = files,
    sample = case_when(
      str_detect(type, "lastz") ~ paste0("GCF_", parse_sample_lastz(basename(files))),
      str_detect(type, "fasta") ~ paste0("GCF_", parse_sample_fasta(basename(files)))
    ),
    count = if (str_detect(type, "lastz")) {
      map_int(files, count_lastz)
    } else {
      map_int(files, count_fasta)
    },
    source = type
  )
}

# Run for all sources
all_counts <- map2_dfr(paths, names(paths), collect_counts)

# Reshape to wide format
df_summary <- all_counts %>%
  select(sample, source, count) %>%
  pivot_wider(names_from = source, values_from = count, values_fill = 0) %>%
  arrange(sample)

print(df_summary)

write.csv(df_summary, "../data/phyluce/sequence_summary.csv")
```

```{r}
summary(df_summary$v1_fasta)
summary(df_summary$v2_fasta)
```

*\~ end of tutorial III \~*

## UCE Processing for Phylogenomics

A different [tutorial for phyluce](https://phyluce.readthedocs.io/en/latest/daily-use/daily-use-3-uce-processing.html#uce-processing) for processing UCE info and getting a matrix / fasta files that can be used to create a tree.

created contig files (just duplicated the `vX-genome-fasta/` output files above and renamed to `vX-contigs`)

```{bash}
cp -r ../data/phyluce/v1-genome-fasta/ ../data/phyluce/v1-contigs/
cp -r ../data/phyluce/v2-genome-fasta/ ../data/phyluce/v2-contigs/
```

```{bash}
# make a directory for log files
mkdir ../data/phyluce/log

# match contigs to probes
cd ../data/phyluce/

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_assembly_match_contigs_to_probes \
    --contigs v1-contigs \
    --probes Baits_Bivalve_v1.fasta \
    --output v1-contigs_to_probes \
    --log-path log
```

```{bash}
# match contigs to probes
cd ../data/phyluce/

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_assembly_match_contigs_to_probes \
    --contigs v2-contigs \
    --probes Baits_Bivalve_v2.fasta \
    --output v2-contigs_to_probes \
    --log-path log
```

### **Creating complete taxon set**

this will give us a list of UCEs that are "complete" within our group of taxa -- meaning present in all taxa.

need to create configuration file with taxon names:

```{bash}
cd ../data/phyluce/v1-contigs/

# write the header first
echo "[dataset1]" > ../taxonset.conf

# append the taxa names (stripping extensions)
for f in *; do
    echo "${f%.*}" >> ../taxonset.conf
done
```

```{bash}
# create the output directory for this taxon set
mkdir ../data/phyluce/taxon-set1/

cd ../data/phyluce/

# create the data matrix configuration file
/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_assembly_get_match_counts \
    --locus-db v1-contigs_to_probes/probe.matches.sqlite \
    --taxon-list-config taxonset.conf \
    --taxon-group 'dataset1' \
    --output dataset1_v1.conf \
    2> dataset1_v1.log
```

```{bash}
cd ../data/phyluce/

# create the data matrix configuration file
/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_assembly_get_match_counts \
    --locus-db v2-contigs_to_probes/probe.matches.sqlite \
    --taxon-list-config taxonset.conf \
    --taxon-group 'dataset1' \
    --output dataset1_v2.conf \
    2> dataset1_v2.log
```

### **Incomplete data matrix**

```{bash}
cd ../data/phyluce/

# create the data matrix configuration file
/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_assembly_get_match_counts \
    --locus-db v1-contigs_to_probes/probe.matches.sqlite \
    --taxon-list-config taxonset.conf \
    --taxon-group 'dataset1' \
    --output dataset1_v1_incomplete.conf \
    --incomplete-matrix \
    2> dataset1_v1.log
```

```{bash}
cd ../data/phyluce/

# create the data matrix configuration file
/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_assembly_get_match_counts \
    --locus-db v2-contigs_to_probes/probe.matches.sqlite \
    --taxon-list-config taxonset.conf \
    --taxon-group 'dataset1' \
    --output dataset1_v2_incomplete.conf \
    --incomplete-matrix \
    2> dataset1_v2.log
```

### **Extracting FASTAs**

complete data matrix

```{bash}
cd ../data/phyluce/

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_assembly_get_fastas_from_match_counts \
    --contigs v1-contigs \
    --locus-db v1-contigs_to_probes/probe.matches.sqlite \
    --match-count-output dataset1_v1.conf \
    --output taxon-set1/dataset1_complete_v1.fasta
```

```{bash}
cd ../data/phyluce/

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_assembly_get_fastas_from_match_counts \
    --contigs v2-contigs \
    --locus-db v2-contigs_to_probes/probe.matches.sqlite \
    --match-count-output dataset1_v2.conf \
    --output taxon-set1/dataset1_complete_v2.fasta
```

preview outputs for V1 and V2

```{bash}
head ../data/phyluce/taxon-set1/dataset1_complete_v1.fasta

echo "number of sequences"
grep -c "^>" ../data/phyluce/taxon-set1/dataset1_complete_v1.fasta
```

```{bash}
head ../data/phyluce/taxon-set1/dataset1_complete_v2.fasta

echo "number of sequences"
grep -c "^>" ../data/phyluce/taxon-set1/dataset1_complete_v2.fasta
```

incomplete data matrix

```{bash}
cd ../data/phyluce/

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_assembly_get_fastas_from_match_counts \
    --contigs v1-contigs \
    --locus-db v1-contigs_to_probes/probe.matches.sqlite \
    --match-count-output dataset1_v1_incomplete.conf \
    --incomplete-matrix dataset1_v1.incomplete \
    --output taxon-set1/dataset1_incomplete_v1.fasta
```

```{bash}
cd ../data/phyluce/

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_assembly_get_fastas_from_match_counts \
    --contigs v2-contigs \
    --locus-db v2-contigs_to_probes/probe.matches.sqlite \
    --match-count-output dataset1_v2_incomplete.conf \
    --incomplete-matrix dataset1_v2.incomplete \
    --output taxon-set1/dataset1_incomplete_v2.fasta
```

```{bash}
head ../data/phyluce/taxon-set1/dataset1_incomplete_v1.fasta

echo "number of sequences"
grep -c "^>" ../data/phyluce/taxon-set1/dataset1_incomplete_v1.fasta
```

```{bash}
head ../data/phyluce/taxon-set1/dataset1_incomplete_v2.fasta

echo "number of sequences"
grep -c "^>" ../data/phyluce/taxon-set1/dataset1_incomplete_v2.fasta
```

now that I have uce datasets, I am switching to the [Phyluce Tutorial I,](https://phyluce.readthedocs.io/en/latest/tutorials/tutorial-1.html#aligning-uce-loci) starting with the 'aligning UCE loci' step, as this will prep the uce datasets and loci for use in IQ tree.

## Align and Trim UCE Fastas (Tutorial I)

we will be using mafft as our aligner. I'll do edge trimming as it is quick, but we will focus on the internal trimming and gblocks conversion when moving on for downstream analysis as it is [more recommended](https://phyluce.readthedocs.io/en/latest/tutorials/tutorial-1.html#aligning-uce-loci) for taxa that may have a wide range of divergence (\>50 MYA), which our bivalves [may have](https://www.sciencedirect.com/science/article/pii/S1055790310003660#f0050).

#### edge trimming

edge trimming for complete v1

```{bash}
cd ../data/phyluce/taxon-set1

mkdir log-cmp_v1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_seqcap_align \
    --input dataset1_complete_v1.fasta \
    --output mafft-nexus-edge-trimmed-cmp_v1 \
    --taxa 18 \
    --aligner mafft \
    --cores 12 \
    --incomplete-matrix \
    --log-path log-cmp_v1

```

summary stats for complete v1 trimming

```{bash}

cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_get_align_summary_data \
    --alignments mafft-nexus-edge-trimmed-cmp_v1 \
    --cores 12 \
    --log-path log-cmp_v1
```

edge trimming for complete v2

```{bash}
cd ../data/phyluce/taxon-set1

mkdir log-cmp_v2

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_seqcap_align \
    --input dataset1_complete_v2.fasta \
    --output mafft-nexus-edge-trimmed-cmp_v2 \
    --taxa 18 \
    --aligner mafft \
    --cores 12 \
    --incomplete-matrix \
    --log-path log-cmp_v2

```

summary stats for complete v2 trimming

```{bash}

cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_get_align_summary_data \
    --alignments mafft-nexus-edge-trimmed-cmp_v2 \
    --cores 12 \
    --log-path log-cmp_v2
```

edge trimming incomplete v1

```{bash}
cd ../data/phyluce/taxon-set1

mkdir log-inc_v1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_seqcap_align \
    --input dataset1_incomplete_v1.fasta \
    --output mafft-nexus-edge-trimmed-inc_v1 \
    --taxa 18 \
    --aligner mafft \
    --cores 12 \
    --incomplete-matrix \
    --log-path log-inc_v1

```

summary stats for incomplete v1

```{bash}

cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_get_align_summary_data \
    --alignments mafft-nexus-edge-trimmed-inc_v1 \
    --cores 12 \
    --log-path log-inc_v1
```

trimming incomplete v2

```{bash}
cd ../data/phyluce/taxon-set1

mkdir log-inc_v2

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_seqcap_align \
    --input dataset1_incomplete_v2.fasta \
    --output mafft-nexus-edge-trimmed-inc_v2 \
    --taxa 18 \
    --aligner mafft \
    --cores 12 \
    --incomplete-matrix \
    --log-path log-inc_v2
```

summary stats for incomplete v2

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_get_align_summary_data \
    --alignments mafft-nexus-edge-trimmed-inc_v2 \
    --cores 12 \
    --log-path log-inc_v2
```

#### **alignment pre internal trimming (fasta output)**

complete v1

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_seqcap_align \
    --input dataset1_complete_v1.fasta \
    --output mafft-nexus-internal-trimmed-cmp_v1 \
    --taxa 18 \
    --aligner mafft \
    --cores 12 \
    --incomplete-matrix \
    --output-format fasta \
    --no-trim \
    --log-path log-cmp_v1
```

complete v2

```{bash}

cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_seqcap_align \
    --input dataset1_complete_v2.fasta \
    --output mafft-nexus-internal-trimmed-cmp_v2 \
    --taxa 18 \
    --aligner mafft \
    --cores 12 \
    --incomplete-matrix \
    --output-format fasta \
    --no-trim \
    --log-path log-cmp_v2

```

incomplete v1

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_seqcap_align \
    --input dataset1_incomplete_v1.fasta \
    --output mafft-nexus-internal-trimmed-inc_v1 \
    --taxa 18 \
    --aligner mafft \
    --cores 12 \
    --incomplete-matrix \
    --output-format fasta \
    --no-trim \
    --log-path log-inc_v1
```

incomplete v2

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_seqcap_align \
    --input dataset1_incomplete_v2.fasta \
    --output mafft-nexus-internal-trimmed-inc_v2 \
    --taxa 18 \
    --aligner mafft \
    --cores 12 \
    --incomplete-matrix \
    --output-format fasta \
    --no-trim \
    --log-path log-inc_v2
```

#### **Gblocks**

takes the aligned fasta ouputs and passes them to parallel wrapper around Gblocks for internal trimming

complete v1

```{bash}

cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_get_gblocks_trimmed_alignments_from_untrimmed \
    --alignments mafft-nexus-internal-trimmed-cmp_v1 \
    --output mafft-nexus-internal-trimmed-gblocks-cmp_v1 \
    --cores 12 \
    --log log-cmp_v1
```

summary stats

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_get_align_summary_data \
    --alignments mafft-nexus-internal-trimmed-gblocks-cmp_v1 \
    --cores 12 \
    --log-path log-cmp_v1
```

complete v2

```{bash}

cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_get_gblocks_trimmed_alignments_from_untrimmed \
    --alignments mafft-nexus-internal-trimmed-cmp_v2 \
    --output mafft-nexus-internal-trimmed-gblocks-cmp_v2 \
    --cores 12 \
    --log log-cmp_v2
```

summary stats

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_get_align_summary_data \
    --alignments mafft-nexus-internal-trimmed-gblocks-cmp_v2 \
    --cores 12 \
    --log-path log-cmp_v2
```

incomplete v1

```{bash}

cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_get_gblocks_trimmed_alignments_from_untrimmed \
    --alignments mafft-nexus-internal-trimmed-inc_v1 \
    --output mafft-nexus-internal-trimmed-gblocks-inc_v1 \
    --cores 12 \
    --log log-inc_v1
```

summary stats

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_get_align_summary_data \
    --alignments mafft-nexus-internal-trimmed-gblocks-inc_v1 \
    --cores 12 \
    --log-path log-inc_v1
```

incomplete v2

```{bash}

cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_get_gblocks_trimmed_alignments_from_untrimmed \
    --alignments mafft-nexus-internal-trimmed-inc_v2 \
    --output mafft-nexus-internal-trimmed-gblocks-inc_v2 \
    --cores 12 \
    --log log-inc_v2
```

summary stats

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_get_align_summary_data \
    --alignments mafft-nexus-internal-trimmed-gblocks-inc_v2 \
    --cores 12 \
    --log-path log-inc_v2
```

#### Alignment Cleaning

removes taxon and locus names from the nexus files and outputs fasta

complete v1

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_remove_locus_name_from_files \
    --alignments mafft-nexus-internal-trimmed-gblocks-cmp_v1 \
    --output mafft-nexus-internal-trimmed-gblocks-clean-cmp_v1 \
    --cores 12 \
    --log-path log-cmp_v1
```

complete v2

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_remove_locus_name_from_files \
    --alignments mafft-nexus-internal-trimmed-gblocks-cmp_v2 \
    --output mafft-nexus-internal-trimmed-gblocks-clean-cmp_v2 \
    --cores 12 \
    --log-path log-cmp_v2
```

incomplete v1

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_remove_locus_name_from_files \
    --alignments mafft-nexus-internal-trimmed-gblocks-inc_v1 \
    --output mafft-nexus-internal-trimmed-gblocks-clean-inc_v1 \
    --cores 12 \
    --log-path log-inc_v1
```

incomplete v2

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_remove_locus_name_from_files \
    --alignments mafft-nexus-internal-trimmed-gblocks-inc_v2 \
    --output mafft-nexus-internal-trimmed-gblocks-clean-inc_v2 \
    --cores 12 \
    --log-path log-inc_v2
```

### Create Data Matrices

Want to get a matrices where 50%, 75%, and 95% of taxa are contained within the alignments. To do this, we will only use the alignments that came from 'incomplete' uce processesing.

#### 50% taxa

incomplete v1

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_get_only_loci_with_min_taxa \
    --alignments mafft-nexus-internal-trimmed-gblocks-clean-inc_v1 \
    --taxa 18 \
    --percent 0.50 \
    --output mafft-nexus-internal-trimmed-gblocks-clean-50p_v1 \
    --cores 12 \
    --log-path log-inc_v1

```

incomplete v2

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_get_only_loci_with_min_taxa \
    --alignments mafft-nexus-internal-trimmed-gblocks-clean-inc_v2 \
    --taxa 18 \
    --percent 0.50 \
    --output mafft-nexus-internal-trimmed-gblocks-clean-50p_v2 \
    --cores 12 \
    --log-path log-inc_v2

```

#### 75% taxa

incomplete v1

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_get_only_loci_with_min_taxa \
    --alignments mafft-nexus-internal-trimmed-gblocks-clean-inc_v1 \
    --taxa 18 \
    --percent 0.75 \
    --output mafft-nexus-internal-trimmed-gblocks-clean-75p_v1 \
    --cores 12 \
    --log-path log-inc_v1

```

incomplete v2

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_get_only_loci_with_min_taxa \
    --alignments mafft-nexus-internal-trimmed-gblocks-clean-inc_v2 \
    --taxa 18 \
    --percent 0.75 \
    --output mafft-nexus-internal-trimmed-gblocks-clean-75p_v2 \
    --cores 12 \
    --log-path log-inc_v2

```

#### 95% taxa

incomplete v1

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_get_only_loci_with_min_taxa \
    --alignments mafft-nexus-internal-trimmed-gblocks-clean-inc_v1 \
    --taxa 18 \
    --percent 0.95 \
    --output mafft-nexus-internal-trimmed-gblocks-clean-95p_v1 \
    --cores 12 \
    --log-path log-inc_v1

```

incomplete v2

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_get_only_loci_with_min_taxa \
    --alignments mafft-nexus-internal-trimmed-gblocks-clean-inc_v2 \
    --taxa 18 \
    --percent 0.95 \
    --output mafft-nexus-internal-trimmed-gblocks-clean-95p_v2 \
    --cores 12 \
    --log-path log-inc_v2

```

### Prep for IQ Tree

need to convert to phylip formal file, which is what IQ tree takes

**50% matrix**

incomplete v1

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_concatenate_alignments \
    --alignments mafft-nexus-internal-trimmed-gblocks-clean-50p_v1 \
    --output mafft-nexus-internal-trimmed-gblocks-clean-50p_v1-raxml \
    --phylip \
    --log-path log-inc_v1
```

incomplete v2

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_concatenate_alignments \
    --alignments mafft-nexus-internal-trimmed-gblocks-clean-50p_v2 \
    --output mafft-nexus-internal-trimmed-gblocks-clean-50p_v2-raxml \
    --phylip \
    --log-path log-inc_v2
```

**75% matrix**

incomplete v1

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_concatenate_alignments \
    --alignments mafft-nexus-internal-trimmed-gblocks-clean-75p_v1 \
    --output mafft-nexus-internal-trimmed-gblocks-clean-75p_v1-raxml \
    --phylip \
    --log-path log-inc_v1
```

incomplete v2

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_concatenate_alignments \
    --alignments mafft-nexus-internal-trimmed-gblocks-clean-75p_v2 \
    --output mafft-nexus-internal-trimmed-gblocks-clean-75p_v2-raxml \
    --phylip \
    --log-path log-inc_v2
```

**95% matrix**

incomplete v1

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_concatenate_alignments \
    --alignments mafft-nexus-internal-trimmed-gblocks-clean-95p_v1 \
    --output mafft-nexus-internal-trimmed-gblocks-clean-95p_v1-raxml \
    --phylip \
    --log-path log-inc_v1
```

incomplete v2

```{bash}
cd ../data/phyluce/taxon-set1

/home/shared/8TB_HDD_02/mewing0/miniforge3/bin/conda run -n phyluce \
/home/shared/8TB_HDD_02/mewing0/miniforge3/envs/phyluce-1.7.3/bin/phyluce_align_concatenate_alignments \
    --alignments mafft-nexus-internal-trimmed-gblocks-clean-95p_v2 \
    --output mafft-nexus-internal-trimmed-gblocks-clean-95p_v2-raxml \
    --phylip \
    --log-path log-inc_v2
```

Can now export the `-raxml` folders with the phylip files for each data matrix and use the phylip files in IQ tree to determine tree model and create tree files!
